Splitted tests on files
