+++ ydb/tests/tools/fq_runner/kikimr_runner.py
@@ -52,7 +52,7 @@ class BaseTenant(abc.ABC):
 
     def stop(self):
         if self.kikimr_cluster:
-            self.kikimr_cluster.stop(kill=True)
+            self.kikimr_cluster.stop(kill=False)
 
     def endpoint(self, node_index=None):
         return "localhost:{}".format(
